pyproject.toml
[tool.poetry]
name = "wps_dabangapi"
version = "0.1.0"
description = ""
authors = ["moonpeter <moonpeter@kakao.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
djangorestframework = "^3.11.0"
notebook = "^6.0.3"
gunicorn = "^20.0.4"
supervisor = "^4.1.0"
requests = "^2.23.0"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
